Move exprtk source file to fields dir.
