+++ unit_tests/fields/new_parsers_test.cpp
@@ -20,7 +20,7 @@
 #include "system/sys_profiler.hh"
 #include "system/file_path.hh"
 
-#include "_parsers/exprtk.hpp"
+#include "fields/exprtk.hpp"
 
 using namespace std;
